Some WKWebViewPrivate `_do*` SPI methods do not compile in Swift https://bugs.webkit.org/show_bug.cgi?id=290084 rdar://147467046
Reviewed by NOBODY (OOPS!).

Some SPI methods such as `_doAfterNextPresentationUpdate` get translated into Swift as `_do(afterNextPresentationUpdate updateBlock:)`.
Not only is this incredibly confusing to read (since the callsite would look like

```
webView._do {
	// stuff
}
```

it is also confusing to the compiler because it is unable to disambiguate the mutliple `_do` methods, and therefore prevents Swift clients
from using these functions.

Fix by manually specifying a Swift name. This has no impact to existing clients, because there aren't any because of the aforementioned
compilation error.

* Source/WebKit/UIProcess/API/Cocoa/WKWebViewPrivate.h: