[Swift in WebKit] Make it easier to use the TestWebKitAPI XC test plan https://bugs.webkit.org/show_bug.cgi?id=294513 rdar://153429368
Reviewed by NOBODY (OOPS!).

Make a few configuration changes to make it easier and more ergonomic to run XC tests:

- Currently, files that are used by both gtests and xctests must be part of both TestWebKitAPILibrary
and TestWebKitAPIBundle. This isn't great, since that means each of these files has to be compiled twice.
To fix this, link TestWebKitAPILibrary from TestWebKitAPIBundle. Ideally, TestWebKitAPILibrary would not
even exist anymore, but that task is significantly more complex.

- Disable code coverage for the xc tests, since it substantially increases the time it takes to run the
tests, and also doesn't provide that much valuable information.

- Move the test plan to the "Everything Up To WebKit + Tools" scheme, so that the same scheme can be used
to both build WebKit and run the tests.

* Tools/TestWebKitAPI/TestBundle/TestWebKitAPI.xctestplan:
* Tools/TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* WebKit.xcworkspace/xcshareddata/xcschemes/Everything up to WebKit + Tools.xcscheme: