[Swift in WebKit] Enable Swift in WTF Need the bug URL (OOPS!). Include a Radar link (OOPS!).
Reviewed by NOBODY (OOPS!).

Enable Swift and Swift-Cxx interop in WTF. In addition to enabling the possibility of using
Swift directly in WTF, it serves two important immediate-term purposes:

1. The module verifier tool does not currently support libraries, such as WTF and bmalloc. Compiling a
Swift file in WTF itself allows for at least a minimal means of verification (specifically, successful
compilation of this file indicates a necessary but not sufficient condition).

2. Enabling interop in WTF is an effective stepping stone to more complex projects like JSC
and WebCore, particularly because it is lower on the WebKit stack.

Notably, doing so uncovered several existing issues in the existing WTF module, which this fixes.

* Source/WTF/Configurations/WTF.xcconfig:

- Define `SCRIPT_OUTPUT_DIR` to be the destination of where the module map will be. More details on this
will follow in the below comments.

- Set `USE_HEADERMAP` to `NO`; it's generally a good idea to have header maps enabled, however this
causes issues with how the module map resolves the files within it, as it causes duplicate definition errors
between the source location and the build products location.

- Define `USER_HEADER_SEARCH_PATHS` so that `config.h` can be included as-is without having to specify a relative
path to the file each time.

- Enable Swift.

* Source/WTF/WTF.xcodeproj/project.pbxproj:

- Remove the script phase that generated the module map now that it is no longer needed.

- Previously, the WTF module map was generated through a Python script which installed the module map directly
in the built products location. Now that the module map is a proper normal file in the source directory, add a
build phase to copy and install it in the built products location.

* Source/WTF/wtf/AutodrainedPool.h:

This file should be able to be included in all environments per module rules, so just compile out the actual
contents of it without producing an error.

* Source/WTF/wtf/RefTrackerMixin.cpp:
* Source/WTF/wtf/RefTrackerMixin.h:
* Source/WTF/wtf/SIMDUTF.cpp:
* Source/WTF/wtf/cocoa/AutodrainedPool.cpp:
* Source/WTF/wtf/cocoa/CrashReporter.cpp:
* Source/WTF/wtf/cocoa/SystemTracingCocoa.cpp:
* Source/WTF/wtf/cocoa/UUIDCocoa.mm:
* Source/WTF/wtf/posix/FileHandlePOSIX.cpp:
* Source/WTF/wtf/posix/MappedFileDataPOSIX.cpp:
* Source/WTF/wtf/spi/cf/CFPrivSPI.h:
* Source/WTF/wtf/spi/cf/CFRunLoopSPI.h:
* Source/WTF/wtf/spi/cf/CFStringSPI.h:
* Source/WTF/wtf/spi/darwin/ProcessMemoryFootprint.h:
* Source/WTF/wtf/text/cocoa/StringImplCocoa.mm:

- The vast majority of wtf properly includes it's own headers using angle bracket includes
instead of quoted includes, but there are a handful of files that do not do this, so fix those.

- There are also some Objective-C header files missing compatibility with C, so fix those too.

* Source/WTF/wtf/wtf.modulemap: Added.

Now for the interesting part!

The existing generated module map had a few issues that prevented it from working properly:

- Some sub-modules were inconsistently and erroneously separated organizationally from the others.
- Some inferred sub-modules were incorrectly requiring C++ when they did not actually need it.
- This was generated in a build phase, which means it is absent in the source directory and was being
generated too late for Swift code within WTF itself to be able to use it, causing compilation issues.

There were also some non-breaking issues:

- Canonically, library modules should be located in the source directory with the filename being
"<library>.modulemap", in this case "wtf.modulemap". This also makes it substantially easier to actually
locate and comprehend the module map, and greatly reduces build complexity.

- Library modules should also be organized somewhat thematically, potentially using umbrellas. The current
approach did not facilitate this; instead, it separated sub-modules based on ad-hoc issues and created them
manually using a script.

To fix and address these issues, the `wtf.modulemap` is introduced, and here's how it works:

Line 1: `module wtf [system]`:

This defines a top-level module named `wtf`. Notably, this does not have the `framework` specifier since `wtf` is just a library. It has the `system` annotation applied since it will be part of the SDK.

Lines 2-14: `explicit module System`:

This defines an explicit sub-module of `wtf` named `System`, which thematically groups the set of headers that define platform or OS specific macros and behaviors. The `explicit` qualifier means that clients need to specify `wtf.Cxx` directly.

Line 15: `export *`

This exports all the headers in the module, including any transitive headers they themselves depend on.

Line 18: `explicit module SPI`

This defines an explicit sub-module named `SPI` which corresponds to the set of headers that forward declare system SPIs. Since these are Objective-C interfaces, they do not require C++.

Line 19: `umbrella "spi"`

This specifies the umbrella directory for the module, in this case the `spi` subdirectory of `wtf`. This means that the module is responsible for all headers in that directory.

Line 21: `explicit module *`

This creates inferred explicit submodules, one for each header in the umbrella directory (equivalent to defining each submodule manually).

Line 26: `explicit module Cxx`

This is the main sub-module for `wtf`; it has to be a sub-module since most WTF
headers require C++ and do not have compilation guards for it, so the requirement
is specified in the module itself. The `Cxx` name is used to match the convention
of how C-families are usually named from Swift.

Line 27: `umbrella "."`

This specifies the umbrella directory for the module, which in this case is the entire directory of `wtf`. Note that this implicitly excludes the headers that were previously specified in prior sub-modules.

Line 36: `export *`

Export all sub-modules from this modules.

Line 38: `link "wtf"`

Libraries need to explicitly link themselves in their module maps, unlike frameworks.

* Source/WTF/wtf/wtf.swift: Added.

Create a basic Swift file as minimal validation of the module map.

* Source/WebGPU/WebGPU/CommandEncoder.swift:

This uses the `wtf` module, but was not importing it, so now it is.