[Swift in WebKit] WebGPU should not disable various verifiers and checkers https://bugs.webkit.org/show_bug.cgi?id=302595 rdar://164839737
Reviewed by NOBODY (OOPS!).

Since verifying and checking things is a good thing to do, stop using ` -no-verify-emitted-module-interface` and
`RUN_SWIFT_ABI_CHECKER_TOOL` in WebGPU. These were only used because the WTF module is incorrect (which itself was not caught
because WTF is a library, which do not support module verification).

Since `Assertions.h` in WTF is unique in that it requires either C, or C++20 and greater, fix by wrapping the parts that need
C++20 or higher in the correct compiler guards.

* Source/WTF/wtf/Assertions.h:
* Source/WebGPU/Configurations/WebGPU.xcconfig: